app/requirements.txt
torch
torchvision
pillow
requests
flask
flask-lambda
Werkzeug==2.0.3
qdrant-client